Fixing some signed/unsigned comparison warnings
svn-id: r43570

+++ engines/gob/save/saveload_playtoons.cpp
@@ -93,20 +93,20 @@ bool SaveLoad_Playtoons::GameHandler::load(int16 dataVar, int32 size, int32 offs
 		size = varSize;
 	}
 
-	if (offset < kPropsSize) {
+	if (((uint32) offset) < kPropsSize) {
 		// Properties
 
-		if ((offset + size) > kPropsSize) {
+		if (((uint32) (offset + size)) > kPropsSize) {
 			warning("Wrong index size (%d, %d)", size, offset);
 			return false;
 		}
 
 		_vm->_inter->_variables->copyFrom(dataVar, _props + offset, size);
 
-	} else if (offset < kPropsSize + kIndexSize) {
+	} else if (((uint32) offset) < kPropsSize + kIndexSize) {
 		// Save index
 
-		if (size != kIndexSize) {
+		if (((uint32) size) != kIndexSize) {
 			warning("Wrong index size (%d, %d)", size, offset);
 			return false;
 		}
@@ -180,20 +180,20 @@ bool SaveLoad_Playtoons::GameHandler::save(int16 dataVar, int32 size, int32 offs
 		size = varSize;
 	}
 
-	if (offset < kPropsSize) {
+	if (((uint32) offset) < kPropsSize) {
 		// Properties
 
-		if ((offset + size) > kPropsSize) {
+		if (((uint32) (offset + size)) > kPropsSize) {
 			warning("Wrong index size (%d, %d)", size, offset);
 			return false;
 		}
 
 		_vm->_inter->_variables->copyTo(dataVar, _props + offset, size);
 
-	}  else if (offset < kPropsSize + kIndexSize) {
+	}  else if (((uint32) offset) < kPropsSize + kIndexSize) {
 		// Save index
 
-		if (size != kIndexSize) {
+		if (((uint32) size) != kIndexSize) {
 			warning("Wrong index size (%d, %d)", size, offset);
 			return false;
 		}
